# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Maison de FLOF API
  version: 1.0.0
  description: Source-inventoried OpenAPI contract for every Next.js API route.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
  /api/admin/articles:
    get:
      operationId: getAdminArticles
      summary: GET /api/admin/articles
      description: Source-inventoried API operation.
      x-contract-tier: stable-minimal
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
    post:
      operationId: postAdminArticles
      summary: POST /api/admin/articles
      description: Source-inventoried API operation.
      x-contract-tier: detailed
      responses:
        "200":
          description: Successful response
          content: &a1
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "201":
          description: Resource created
          content: *a1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    patch:
      operationId: patchAdminArticles
      summary: PATCH /api/admin/articles
      description: Source-inventoried API operation.
      x-contract-tier: detailed
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    delete:
      operationId: deleteAdminArticles
      summary: DELETE /api/admin/articles
      description: Source-inventoried API operation.
      x-contract-tier: detailed
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
...
  /api/admin/products:
    get:
      operationId: getAdminProducts
      summary: GET /api/admin/products
      description: Source-inventoried API operation.
      x-contract-tier: stable-minimal
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
    post:
      operationId: postAdminProducts
      summary: POST /api/admin/products
      description: Source-inventoried API operation.
      x-contract-tier: detailed
      responses:
        "200":
          description: Successful response
          content: &a11
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "201":
          description: Resource created
          content: *a11
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    patch:
      operationId: patchAdminProducts
      summary: PATCH /api/admin/products
      description: Source-inventoried API operation.
      x-contract-tier: detailed
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    delete:
      operationId: deleteAdminProducts
      summary: DELETE /api/admin/products
      description: Source-inventoried API operation.
      x-contract-tier: detailed
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required:
        - error
        - requestId
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - TOO_MANY_REQUESTS
                - RATE_LIMITED
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
            message:
              type: string
            details: {}
        requestId:
          type: string
...
    GenericObject:
      type: object
      additionalProperties: true
    GenericArray:
      type: array
      items:
        $ref: "#/components/schemas/GenericObject"
    Success:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          const: true
...